Add meter audit db store
