+++ cmd/portal/analytic/wire_gen.go
@@ -71,17 +71,22 @@ func NewProjectWeeklyReport(ctx context.Context, pool *db.Pool, databaseCredenti
 	readHandle := auditdb.NewReadHandle(ctx, pool, databaseEnvironmentConfig, auditDatabaseCredentials, factory)
 	auditdbSQLBuilder := auditdb.NewSQLBuilder(auditDatabaseCredentials)
 	readSQLExecutor := auditdb.NewReadSQLExecutor(ctx, readHandle)
-	auditDBReadStore := &analytic.AuditDBReadStore{
+	auditDBReadStore := &meter.AuditDBReadStore{
+		SQLBuilder:  auditdbSQLBuilder,
+		SQLExecutor: readSQLExecutor,
+	}
+	analyticAuditDBReadStore := &analytic.AuditDBReadStore{
 		SQLBuilder:  auditdbSQLBuilder,
 		SQLExecutor: readSQLExecutor,
 	}
 	projectWeeklyReport := &analytic.ProjectWeeklyReport{
-		GlobalHandle:  handle,
-		GlobalDBStore: globalDBStore,
-		AppDBHandle:   appdbHandle,
-		AppDBStore:    appDBStore,
-		AuditDBHandle: readHandle,
-		AuditDBStore:  auditDBReadStore,
+		GlobalHandle:      handle,
+		GlobalDBStore:     globalDBStore,
+		AppDBHandle:       appdbHandle,
+		AppDBStore:        appDBStore,
+		AuditDBHandle:     readHandle,
+		MeterAuditDBStore: auditDBReadStore,
+		AuditDBStore:      analyticAuditDBReadStore,
 	}
 	return projectWeeklyReport
 }
@@ -139,6 +144,10 @@ func NewCountCollector(ctx context.Context, pool *db.Pool, databaseCredentials *
 		SQLExecutor: readSQLExecutor,
 	}
 	writeHandle := auditdb.NewWriteHandle(ctx, pool, databaseEnvironmentConfig, auditDatabaseCredentials, factory)
+	meterAuditDBReadStore := &meter.AuditDBReadStore{
+		SQLBuilder:  auditdbSQLBuilder,
+		SQLExecutor: readSQLExecutor,
+	}
 	writeSQLExecutor := auditdb.NewWriteSQLExecutor(ctx, writeHandle)
 	auditDBWriteStore := &analytic.AuditDBWriteStore{
 		SQLBuilder:  auditdbSQLBuilder,
@@ -161,6 +170,7 @@ func NewCountCollector(ctx context.Context, pool *db.Pool, databaseCredentials *
 		AuditDBReadHandle:  readHandle,
 		AuditDBReadStore:   auditDBReadStore,
 		AuditDBWriteHandle: writeHandle,
+		MeterAuditDBStore:  meterAuditDBReadStore,
 		AuditDBWriteStore:  auditDBWriteStore,
 		AnalyticService:    service,
 	}
